scripts/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
starlette==0.36.3
httpx==0.26.0
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
psycopg2-binary